Create Menu Class To Display Options
